requirements.txt
# BirdCLEF+ 2026 — local training / offline pipeline
# Kaggle notebook runtime is managed separately (pinned TF 2.20 wheels).

# Core
numpy>=1.26,<2.1
pandas>=2.1,<3
pyarrow>=15                     # for meta.parquet / filenames.parquet
scipy>=1.11,<2
scikit-learn>=1.4,<2
tqdm>=4.66
pyyaml>=6
soundfile>=0.12
resampy>=0.4

# Perch / TF (CPU on Mac works via pip wheel for arm64)
tensorflow>=2.15,<2.21 ; platform_system == "Darwin"

# Optional boosting for probes
lightgbm>=4.3

# PyTorch for MPS student / CNN head
# Install manually for MPS: pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
# or just: pip install torch torchvision
torch>=2.2
torchvision>=0.17

# Kaggle data sync
kagglehub>=1.0

# Dev
pytest>=8
